requirements.txt
asgiref==3.8.1
autopep8==1.4.4
crispy-bootstrap4==2024.1
Django==4.2.11
django-allauth==0.62.1
django-crispy-forms==2.1
django-debug-toolbar==4.3.0
pep8==1.7.1
pycodestyle==2.5.0
python-decouple==3.1
pytz==2018.5
sqlparse==0.5.0
typing_extensions==4.11.0
tzdata==2024.1
